feat: middleware-related stuff separated from client
The original client.py was divided in two files,
client_base.py, where only NetApp communication is,
and client.py, which inherits from client_base and adds
the middleware-related communication.

The gstreamer part was integrated directly into the
client_base -> in the future, it would be nice to
make gstreamer and websocket communication as a plugin.

Examples adapted to the new client interface and made
more general.

+++ era_5g_client/client.py
@@ -1,14 +1,15 @@
-import base64
+import logging
 import os
 from collections.abc import Callable
+from enum import Enum
 from typing import Dict, List, Optional, Tuple
 
-import cv2
 import numpy as np
 import requests
-import socketio
 from requests import HTTPError, Response
 
+from era_5g_client.client_base import NetAppClientBase
+from era_5g_client.dataclasses import MiddlewareInfo, NetAppLocation
 from era_5g_client.exceptions import FailedToConnect, NetAppNotReady
 from era_5g_client.middleware_resource_checker import MiddlewareResourceChecker
 
@@ -17,32 +18,26 @@ buffer: List[Tuple[np.ndarray, Optional[str]]] = []
 # port of the netapp's server
 NETAPP_PORT = int(os.getenv("NETAPP_PORT", 5896))
 
-"""
-TODO ZdenekM: I would divide this to
-    NetAppClientBase (without middleware),
-    NetAppClient (derived from the above, which will be using middleware)
-"""
+
+class RunTaskMode(Enum):
+    # deploy the task but dont wait until it is ready, do not register with it
+    DO_NOTHING = 1
+    # wait until the netapp is ready, do not register with it
+    WAIT = 2
+    # wait until the netapp is ready and register with it afterwards
+    WAIT_AND_REGISTER = 3
 
 
-class NetAppClient:
-    """Basic implementation of the NetApp client.
+class NetAppClient(NetAppClientBase):
+    """Extension of the NetAppClientBase class, which enable communication with
+    the Middleware.
 
-    It creates the Requests object with session and bind callbacks for
-    connection info and results from the NetApp.
+    It allow to deploy the NetApp and check on the status of the NetApp
     """
 
     def __init__(
         self,
-        host: str,
-        user_id: str,
-        password: str,
-        task_id: str,
-        resource_lock: bool,
         results_event: Callable,
-        use_middleware: bool = True,
-        wait_for_netapp: bool = True,
-        netapp_uri: Optional[str] = None,
-        netapp_port: Optional[int] = None,
         image_error_event: Optional[Callable] = None,
         json_error_event: Optional[Callable] = None,
     ) -> None:
@@ -73,224 +68,139 @@ class NetAppClient:
                 the middleware
         """
 
-        # temporary removed asserts until the middleware will provide necessary info
+        super().__init__(results_event, image_error_event, json_error_event)
 
-        # if middleware is used, the netapp_uri and port should be obtained from it
-        # assert use_middleware and not netapp_uri and not netapp_port
-        # if middleware is not used, the netapp_uri and port needs to be provided
-        # assert not use_middleware and netapp_uri and netapp_port
-
-        self.sio = socketio.Client()
-        self.session = requests.session()
-        self.host = host.rstrip("/") if host is not None else None
-        self.resource_lock = resource_lock
-        self.netapp_host = netapp_uri
-        self.netapp_port = netapp_port
-        self.use_middleware = use_middleware
-        self.sio.on("message", results_event, namespace="/results")
-        self.sio.on("connect", self.on_connect_event, namespace="/results")
-        self.sio.on("image_error", image_error_event, namespace="/data")
-        self.sio.on("json_error", json_error_event, namespace="/data")
-        self.sio.on("connect_error", self.on_connect_error, namespace="/results")
-        self.session_cookie: Optional[str] = None
+        self.host: Optional[str] = None
         self.action_plan_id: Optional[str] = None
         self.resource_checker: Optional[MiddlewareResourceChecker] = None
-        self.user_id = user_id
-        self.password = password
-        self.task_id = task_id
+        self.middleware_info: Optional[MiddlewareInfo] = None
         self.token: Optional[str] = None
-        self.wait_for_netapp = wait_for_netapp
-        self.ws_data: Optional[bool] = False
-        if self.use_middleware:
-            self.connect_to_middleware()
 
-    def test(data):
-        print(data)
+    def connect_to_middleware(self, middleware_info: MiddlewareInfo) -> None:
+        """Authenticates with the middleware and obtains a token for future
+        calls.
 
-    def connect_to_middleware(self) -> None:
+        Args:
+            host (str): The IP address or hostname of the middleware
+            user_id (str): GUID of the middleware user
+            password (str): The password of the middleware user
 
+        Raises:
+            FailedToConnect: Raised when the authentication with the
+                middleware failed
+        """
+        self.middleware_info = middleware_info
+        self.middleware_info.uri = self.middleware_info.uri.rstrip("/")
         try:
             # connect to the middleware
-            self.token = self.gateway_login(self.user_id, self.password)
+            self.token = self.gateway_login(self.middleware_info.user_id, self.middleware_info.password)
+        except FailedToConnect as ex:
+            logging.error(f"Can't connect to middleware: {ex}")
+            raise
 
+    def run_task(
+        self,
+        task_id: str,
+        robot_id: str,
+        resource_lock: bool,
+        mode: Optional[RunTaskMode] = RunTaskMode.WAIT_AND_REGISTER,
+        gstreamer: Optional[bool] = False,
+        ws_data: Optional[bool] = False,
+        args: Optional[Dict] = None,
+    ) -> None:
+        """Deploys the task with provided *task_id* using middleware and
+        (optionally) waits until the netapp is ready and register with it.
+
+        Args:
+            task_id (str): The GUID of the task which should be deployed.
+            robot_id (str): The GUID of the robot that deploys the NetApp.
+            resource_lock (bool): TBA
+            mode (Optional[RunTaskMode]): Specify the mode in which the run_task
+                works
+            gstreamer (Optional[bool], optional):  Indicates if a GStreamer pipeline
+                should be initialized for image transport. Applied only if register
+                is True. Defaults to False.
+            ws_data (Optional[bool], optional): Indicates if a separate websocket channel
+                for data transport should be set. Applied only if register
+                is True. Defaults to False.
+            args (Optional[Dict], optional): NetApp-specific arguments. Applied only if register
+                is True. Defaults to None.
+
+        Raises:
+            FailedToConnect: Raised when running the task failed.
+        """
+        assert self.middleware_info
+        try:
             self.action_plan_id = self.gateway_get_plan(
-                self.task_id, self.resource_lock
+                task_id, resource_lock, robot_id
             )  # Get the plan_id by sending the token and task_id
+
             if not self.action_plan_id:
-                raise FailedToConnect("Need plan id first...")
+                raise FailedToConnect("Failed to obtain action plan id...")
 
             self.resource_checker = MiddlewareResourceChecker(
-                self.token,
+                str(self.token),
                 self.action_plan_id,
-                self.build_middleware_api_endpoint("orchestrate/orchestrate/plan"),
+                self.middleware_info.build_api_endpoint("orchestrate/orchestrate/plan"),
                 daemon=True,
             )
-            self.resource_checker.start()
-            if self.use_middleware and self.wait_for_netapp:
-                self.wait_until_netapp_ready()
-                print(self.resource_checker.is_ready())
-                self.load_netapp_uri()
-            print("Feedback: Got new plan successfully.")
-        except Exception as ex:
-            self.delete_all_resources()
-            raise FailedToConnect(f"Can't connect to middleware: {ex}") from ex
 
-    def register(self, ws_data: Optional[bool] = False, args: Optional[Dict] = None) -> Response:
+            self.resource_checker.start()
+            if mode in [RunTaskMode.WAIT, RunTaskMode.WAIT_AND_REGISTER]:
+                self.wait_until_netapp_ready()
+                self.load_netapp_uri()
+                if not self.netapp_location:
+                    raise FailedToConnect("Failed to obtain NetApp URI or port")
+                if mode == RunTaskMode.WAIT_AND_REGISTER:
+                    self.register(self.netapp_location, gstreamer, ws_data, args)
+        except (FailedToConnect, NetAppNotReady) as ex:
+            self.delete_all_resources()
+            logging.error(f"Failed to run task: {ex}")
+            raise
+
+    def register(
+        self,
+        netapp_location: NetAppLocation,
+        gstreamer: Optional[bool] = False,
+        ws_data: Optional[bool] = False,
+        args: Optional[Dict] = None,
+    ) -> Response:
         """Calls the /register endpoint of the NetApp interface and if the
         registration is successful, it sets up the WebSocket connection for
         results retrieval.
 
         Args:
-            ws_data (bool): indicates if a separate websocket channel
-                for data transport should be set
-            args (dict): optional parameters to be passed to
-            the NetApp, in the form of dict. Defaults to None.
+            netapp_location (NetAppLocation): The URI and port of the NetApp interface.
+            gstreamer (Optional[bool], optional):  Indicates if a GStreamer pipeline
+                should be initialized for image transport. Defaults to False.
+            ws_data (Optional[bool], optional): Indicates if a separate websocket channel
+                for data transport should be set. Defaults to False.
+            args (Optional[Dict], optional): NetApp-specific arguments. Defaults to None.
+
+        Raises:
+            NetAppNotReady: Raised when register called before the NetApp is ready.
+
         Returns:
-            Response: response from the NetApp
+            Response: The response from the NetApp.
         """
 
-        if self.use_middleware:
+        if not self.resource_checker:
+            raise NetAppNotReady("Not connected to the middleware.")
 
-            if not self.resource_checker:
-                raise NetAppNotReady("Not connected to the middleware.")
+        if not self.resource_checker.is_ready():
+            raise NetAppNotReady("Not ready.")
 
-            if not self.resource_checker.is_ready():
-                raise NetAppNotReady("Not ready.")
-
-        self.ws_data = ws_data
-
-        response = self.session.post(
-            self.build_netapp_api_endpoint("register"), json=args, headers={"Content-Type": "application/json"}
-        )
-
-        # checks whether the NetApp responded with any data
-        if len(response.content) > 0:
-            try:
-                data = response.json()
-            except ValueError as ex:
-                raise FailedToConnect(f"Decoding JSON has failed: {ex}, response: {response}")
-            # checks if an error was returned
-            if "error" in data:
-                raise FailedToConnect(f"{response.status_code}: {data['error']}")
-
-        self.session_cookie = response.cookies["session"]
-
-        # creates the WebSocket connection
-        if ws_data:
-            self.sio.connect(
-                self.build_netapp_api_endpoint(""),
-                namespaces=["/results", "/data"],
-                headers={"Cookie": f"session={self.session_cookie}"},
-                wait_timeout=10,
-            )
-        else:
-            self.sio.connect(
-                self.build_netapp_api_endpoint(""),
-                namespaces=["/results"],
-                headers={"Cookie": f"session={self.session_cookie}"},
-                wait_timeout=10,
-            )
+        response = super().register(netapp_location, gstreamer, ws_data, args)
 
         return response
 
     def disconnect(self) -> None:
         """Calls the /unregister endpoint of the server and disconnects the
         WebSocket connection."""
-
-        if self.netapp_host is not None and self.netapp_port is not None:
-            self.session.post(self.build_netapp_api_endpoint("unregister"))
-        self.sio.disconnect()
-        if self.use_middleware:
-            self.delete_all_resources()
-
-    def build_middleware_api_endpoint(self, path: str) -> str:
-        """Builds an API endpoint on the NetApp interface.
-
-        Args:
-            path (str): endpoint path
-        Returns:
-            _type_: complete URI to requested endpoint
-        """
-
-        return f"http://{self.host}/{path}"
-
-    def build_netapp_api_endpoint(self, path: str) -> str:
-        """Builds an API endpoint on the NetApp interface.
-
-        Args:
-            path (str): endpoint path
-        Returns:
-            _type_: complete URI to requested endpoint
-        """
-
-        return f"http://{self.netapp_host}:{self.netapp_port}/{path}"
-
-    def wait(self) -> None:
-        """Blocking infinite waiting."""
-        self.sio.wait()
-
-    def on_connect_event(self) -> None:
-        """The callback called once the connection to the NetApp is made."""
-        print("Connected to server")
-
-    def on_connect_error(self, data: str) -> None:
-        """The callback called on connection error."""
-        print(f"Connection error: {data}")
-        # self.disconnect()
-
-    def send_image_http(self, frame: np.ndarray, timestamp: Optional[str] = None, batch_size: int = 1) -> None:
-        """Encodes the image frame to the jpg format and sends it over the
-        HTTP, to the /image endpoint.
-
-        Args:
-            frame (np.ndarray): image frame
-            timestamp (str, optional): Frame timestamp The timestamp format
-            is defined by the NetApp. Defaults to None.
-            batch_size (int, optional): If higher than one, the images are
-            send in batches. Defaults to 1
-        """
-
-        assert batch_size >= 1
-
-        _, img_encoded = cv2.imencode(".jpg", frame)
-
-        if len(buffer) < batch_size:
-            buffer.append((img_encoded, timestamp))
-
-        if len(buffer) == batch_size:
-            # TODO find out right type annotation
-            files = [("files", (f"image{i + 1}", buffer[i][0], "image/jpeg")) for i in range(batch_size)]
-
-            timestamps: List[Optional[str]] = [b[1] for b in buffer]
-            self.session.post(
-                self.build_netapp_api_endpoint("image"),
-                files=files,  # type: ignore
-                params={"timestamps[]": timestamps},  # type: ignore
-            )
-            buffer.clear()
-
-    def send_image_ws(self, frame: np.ndarray, timestamp: Optional[str] = None):
-        """Encodes the image frame to the jpg format and sends it over the
-        websocket, to the /data namespace.
-
-        Args:
-            frame (np.ndarray): Image frame
-            timestamp (Optional[str], optional): Frame timestamp The timestamp format
-            is defined by the NetApp. Defaults to None.
-        """
-        assert self.ws_data
-
-        _, img_encoded = cv2.imencode(".jpg", frame)
-        self.sio.emit("image", {"timestamp": timestamp, "frame": base64.b64encode(img_encoded)}, "/data")
-
-    def send_json_ws(self, json: dict):
-        """Sends netapp-specific json data.
-
-        Args:
-            json (dict): Json data in the form of Python dictionary
-        """
-        assert self.ws_data
-        self.sio.emit("json", json, "/data")
+        super().disconnect()
+        if self.resource_checker is not None:
+            self.resource_checker.stop()
+        self.delete_all_resources()
 
     def wait_until_netapp_ready(self) -> None:
         """Blocking wait until the NetApp is running.
@@ -299,20 +209,20 @@ class NetAppClient:
             NetAppNotReady: _description_
         """
         if not self.resource_checker:
-            raise NetAppNotReady("Not connected to middleware.")
+            raise FailedToConnect("Not connected to middleware.")
         self.resource_checker.wait_until_resource_ready()
 
     def load_netapp_uri(self) -> None:
         if not (self.resource_checker and self.resource_checker.is_ready()):
             raise NetAppNotReady
-        self.netapp_host = self.resource_checker.url
-        self.netapp_port = NETAPP_PORT
+        self.netapp_location = NetAppLocation(str(self.resource_checker.url), NETAPP_PORT)
 
     def gateway_login(self, id: str, password: str) -> str:
+        assert self.middleware_info
         print("Trying to log into the middleware")
         # Request Login
         try:
-            r = requests.post(self.build_middleware_api_endpoint("Login"), json={"Id": id, "Password": password})
+            r = requests.post(self.middleware_info.build_api_endpoint("Login"), json={"Id": id, "Password": password})
             response = r.json()
             if "errors" in response:
                 raise FailedToConnect(str(response["errors"]))
@@ -327,14 +237,21 @@ class NetAppClient:
         except KeyError as e:
             raise FailedToConnect(f"Could not login to the middleware gateway, the response does not contain {e}")
 
-    def gateway_get_plan(self, taskid: str, resource_lock: bool) -> str:
+    def gateway_get_plan(self, taskid: str, resource_lock: bool, robot_id: str) -> str:
+        assert self.middleware_info
         # Request plan
+
         try:
             print("Goal task is: " + str(taskid))
             hed = {"Authorization": f"Bearer {str(self.token)}"}
-            data = {"TaskId": str(taskid), "LockResourceReUse": resource_lock}
-            response = requests.post(self.build_middleware_api_endpoint("Task/Plan"), json=data, headers=hed).json()
-
+            data = {
+                "TaskId": str(taskid),
+                "LockResourceReUse": resource_lock,
+                "RobotId": robot_id,
+            }
+            response = requests.post(
+                self.middleware_info.build_api_endpoint("Task/Plan"), json=data, headers=hed
+            ).json()
             if not isinstance(response, dict):
                 raise FailedToConnect("Invalid response.")
 
@@ -342,9 +259,8 @@ class NetAppClient:
                 raise FailedToConnect(f"response {response['statusCode']}: {response['message']}")
             # todo:             if "errors" in response:
             #                 raise FailedToConnect(str(response["errors"]))
-            # print(action_sequence)
             action_plan_id = str(response["ActionPlanId"])
-            print("ActionPlanId ** is: " + str(self.action_plan_id))
+            print("ActionPlanId ** is: " + str(action_plan_id))
             return action_plan_id
         except KeyError as e:
             raise FailedToConnect(f"Could not get the plan: {e}")
@@ -355,11 +271,14 @@ class NetAppClient:
 
         try:
             hed = {"Authorization": "Bearer " + str(self.token)}
-            url = self.build_middleware_api_endpoint(f"orchestrate/orchestrate/plan/{str(self.action_plan_id)}")
-            response = requests.delete(url, headers=hed)
+            if self.middleware_info:
+                url = self.middleware_info.build_api_endpoint(
+                    f"orchestrate/orchestrate/plan/{str(self.action_plan_id)}"
+                )
+                response = requests.delete(url, headers=hed)
 
-            if response.ok:
-                print("Resource deleted")
+                if response.ok:
+                    print("Resource deleted")
 
         except HTTPError as e:
             print(e.response.status_code)
